--- snakemake_executor_plugin_azure_batch/common.py
@@ -1,11 +1,0 @@
-def bytesto(bytes, to, bsize=1024):
-    """convert bytes to megabytes.
-    bytes to mb: bytesto(bytes, 'm')
-    bytes to gb: bytesto(bytes, 'g' etc.
-    From https://gist.github.com/shawnbutts/3906915
-    """
-    levels = {"k": 1, "m": 2, "g": 3, "t": 4, "p": 5, "e": 6}
-    answer = float(bytes)
-    for _ in range(levels[to]):
-        answer = answer / bsize
-    return answer
